remove hackaround for upstream caching bug
THIS IS BREAKING FOR WINDOWS INSTALLATIONS WITH CACHELIB < 0.9.0. a bug in older versions of `cachelib` broke filesystem caching via `flask-caching` on Windows, which we fixed with an ugly hackaround. this bug was fixed in `cachelib` 0.9.0, usage of which has slowly trickled downstream, and is now causing our hackaround to break new Windows installs. this commit removes the hackaround entirely. existing Windows users will need to update their dependencies.