fix some color things and make output fonts configurable
The "none" color option for highlighted points in output plots always
made those points transparent rather than retaining the color option
of the primary scatter trace. Making color scales consistent between
the colorbar, highlight scatter, and main scatter requires use of a
shared Norm object, as well as a bunch of gross logic to decide which portions
of the data to use to construct that object depending on various
combinations of color options etc. So this commit also rearranges a lot of
the logic related to colormap bounds setting and normalization.

This commit also makes some things about matplotlib/plotly colormap
matching more consistent. For cases in which there is no match,
instead of erroring out, it uses 'inferno' as a fallback and raises a
warning.

This commit also makes output font files and font sizes configurable
in plotter.config.user_output_style.

Finally, it deletes all code (i think) related to the old output functionality.